requirements.txt
streamlit
pandas
cohere

python-dotenv # For environment variables
plotly
